requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
torch==2.2.1
open_clip_torch==2.24.0
pdfkit==1.0.0
backoff==2.2.1
google-api-core==2.17.1
numpy==1.24.3
nltk==3.8.1 